hal: Usb: Handle failure cases for readFile during autosuspend check
In the event readFile fails to open file for checking interface class,
the return value is not checked. Instead only data length is checked.
Invoking stoi on this wrong string saw a SIGABRT crash. Initialize
the string to be read to be an empty string so that if fopen fails
the length check ensures we don't access any invalid or corrupt data.

Change-Id: If2c3b081e9de5c4c801d98fe631c1ac997a2d9ba